+++ tests/testthat/test-pick_sl.R
@@ -1,0 +1,3 @@
+test_that("pick_sl", {
+  expect_equal(length(pick_sl(10, verbose = FALSE)[[1]]), 7)
+})